move static methods from cli.py and model.py to ctimer/utils.py so they can be used by tests
